requirements.txt
Flask==2.3.3
celery==5.3.4
redis==4.6.0
qrcode==7.4.2
Pillow==10.0.0
requests==2.31.0
Flask-SQLAlchemy==3.0.5